Add capture-vmcore tool
This tool can be used to capture vmcore from VMs that are paused at a
kernel crash sink. This vmcore is similar to what a kdump kernel would
present under /proc/vmcore and is compatible with all associated tools
such as makedumpfile, crash, etc.

The VM kernel just needs to have an arbitrary kdump
kernel loaded. However we do not depend on that kernel. The kexec
doesn't need to be successful and the tool doesn't need a second .json
corredponding to the kdump kernel.