Give VM forks friendly and identifiable names
Uses libxenlight (new dependency) to rename sink and fuzzing forks.
Names are derived from the (grand)parent VM but a user specified token
(-G) will also be mixed in if specified. The usecase for -G is to be
able to easily idendify VM forks when triaging a set of crashing
inputs with --keep.
fork's domid is used in place if -G is not specified.

Example VM names:

For kfx without -G:
debian-1-sink-2
debian-1-fuzz-3

For kfx with -G id:000000:
debian-1-sink-id:000000
debian-1-fuzz-id:000000

For forkvm:
debian-1-forkvm-2